# docs/specs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Stoneboyz CRM API
  version: 0.1.0
  description: |
    Source of truth for all API contracts.
    Edit this file before implementing any endpoint.
    packages/api-client is generated from this file; do not hand-edit it.
servers:
  - url: http://localhost:3001/api/v1
    description: Local development
  - url: https://api.stoneboyz.app/v1
    description: Production placeholder
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    cookieAuth:
      type: apiKey
      in: cookie
      name: better-auth.session_token
  parameters:
    CustomerId:
      name: customerId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    QuoteId:
      name: quoteId
      in: path
      required: true
      description: UUID of the quote
      schema:
        type: string
        format: uuid
...
    AreaId:
      name: areaId
      in: path
      required: true
      description: UUID of the quote area
      schema:
        type: string
        format: uuid
    PieceId:
      name: id
      in: path
      required: true
      description: UUID of the counter piece
      schema:
        type: string
        format: uuid
    EdgeId:
      name: id
      in: path
      required: true
      description: UUID of the edge segment
      schema:
        type: string
        format: uuid
    SinkId:
      name: id
      in: path
      required: true
      description: UUID of the sink cutout
      schema:
        type: string
        format: uuid
...
  schemas:
    ErrorResponse:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          example: VALIDATION_ERROR
        message:
          type: string
          example: Request validation failed
        details:
          type: object
...
    ActorRequest:
      type: object
      properties:
...
    QuoteArea:
      type: object
      required:
        - id
        - quoteId
        - sortOrder
        - name
        - subtotalCents
        - measurementTotals
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
          format: uuid
        quoteId:
          type: string
          format: uuid
        sortOrder:
          type: integer
          default: 0
        name:
          type: string
        material:
          type: string
          nullable: true
        color:
          type: string
          nullable: true
        edgeProfile:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        subtotalCents:
          type: integer
        measurementTotals:
          $ref: '#/components/schemas/QuoteMeasurementAreaTotals'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    QuoteMeasurementAreaTotals:
      type: object
      required: [pieceCount, countertopSqFt, finishedEdgeLinFt, splashSqFt, sinkCutoutCount, faucetHoleCount]
      properties:
        pieceCount:
          type: integer
        countertopSqFt:
          type: number
        finishedEdgeLinFt:
          type: number
        splashSqFt:
          type: number
        sinkCutoutCount:
          type: integer
        faucetHoleCount:
          type: integer
...
    CounterPiece:
      type: object
      required: [id, quoteAreaId, sortOrder, name, lengthIn, widthIn, quantity, createdAt, updatedAt]
      properties:
        id:
          type: string
          format: uuid
        quoteAreaId:
          type: string
          format: uuid
        sortOrder:
          type: integer
          default: 0
        name:
          type: string
          nullable: true
        lengthIn:
          type: number
          exclusiveMinimum: 0
        widthIn:
          type: number
          exclusiveMinimum: 0
        quantity:
          type: integer
          minimum: 1
          default: 1
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CreateCounterPieceRequest:
      type: object
      required: [lengthIn, widthIn]
      properties:
        sortOrder:
          type: integer
          default: 0
        name:
          type: string
          minLength: 1
        lengthIn:
          type: number
          exclusiveMinimum: 0
        widthIn:
          type: number
          exclusiveMinimum: 0
        quantity:
          type: integer
          minimum: 1
          default: 1
    UpdateCounterPieceRequest:
      type: object
      description: At least one updatable field is required.
      minProperties: 1
      properties:
        sortOrder:
          type: integer
        name:
          type: string
          nullable: true
        lengthIn:
          type: number
          exclusiveMinimum: 0
        widthIn:
          type: number
          exclusiveMinimum: 0
        quantity:
          type: integer
          minimum: 1
...
    EdgeSegment:
      type: object
      required: [id, quoteAreaId, sortOrder, lengthIn, treatment, splashHeightIn, createdAt, updatedAt]
      properties:
        id:
          type: string
          format: uuid
        quoteAreaId:
          type: string
          format: uuid
        sortOrder:
          type: integer
          default: 0
        lengthIn:
          type: number
          exclusiveMinimum: 0
        treatment:
          $ref: '#/components/schemas/EdgeTreatment'
        splashHeightIn:
          type: number
          exclusiveMinimum: 0
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CreateEdgeSegmentRequest:
      type: object
      required: [lengthIn, treatment]
      properties:
        sortOrder:
          type: integer
          default: 0
        lengthIn:
          type: number
          exclusiveMinimum: 0
        treatment:
          $ref: '#/components/schemas/EdgeTreatment'
        splashHeightIn:
          type: number
          exclusiveMinimum: 0
    UpdateEdgeSegmentRequest:
      type: object
      description: At least one updatable field is required.
      minProperties: 1
      properties:
        sortOrder:
          type: integer
        lengthIn:
          type: number
          exclusiveMinimum: 0
        treatment:
          $ref: '#/components/schemas/EdgeTreatment'
        splashHeightIn:
          type: number
          exclusiveMinimum: 0
          nullable: true
...
    SinkCutout:
      type: object
      required: [id, quoteAreaId, sortOrder, quantity, model, sinkType, shape, cutoutLengthIn, cutoutWidthIn, faucetHoleCount, centerline, createdAt, updatedAt]
      properties:
        id:
          type: string
          format: uuid
        quoteAreaId:
          type: string
          format: uuid
        sortOrder:
          type: integer
          default: 0
        quantity:
          type: integer
          minimum: 1
          default: 1
        model:
          type: string
          nullable: true
        sinkType:
          $ref: '#/components/schemas/SinkType'
        shape:
          $ref: '#/components/schemas/SinkShape'
        cutoutLengthIn:
          type: number
          exclusiveMinimum: 0
        cutoutWidthIn:
          type: number
          exclusiveMinimum: 0
        faucetHoleCount:
          type: integer
          minimum: 0
          maximum: 5
          default: 0
        centerline:
          $ref: '#/components/schemas/SinkCenterline'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CreateSinkCutoutRequest:
      type: object
      required: [sinkType, shape, cutoutLengthIn, cutoutWidthIn]
      properties:
        sortOrder:
          type: integer
          default: 0
        quantity:
          type: integer
          minimum: 1
          default: 1
        model:
          type: string
          minLength: 1
        sinkType:
          $ref: '#/components/schemas/SinkType'
        shape:
          $ref: '#/components/schemas/SinkShape'
        cutoutLengthIn:
          type: number
          exclusiveMinimum: 0
        cutoutWidthIn:
          type: number
          exclusiveMinimum: 0
        faucetHoleCount:
          type: integer
          minimum: 0
          maximum: 5
          default: 0
        centerline:
          $ref: '#/components/schemas/SinkCenterline'
...
    UpdateQuoteAreaRequest:
      type: object
      description: At least one updatable field is required.
      minProperties: 2
      properties:
        name:
          type: string
          minLength: 1
        sortOrder:
          type: integer
        material:
          type: string
          nullable: true
        color:
          type: string
          nullable: true
        edgeProfile:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
...
security:
  - bearerAuth: []
paths:
...
  /customers/{customerId}/quotes/{quoteId}/areas/{areaId}:
    patch:
      operationId: updateQuoteArea
      summary: Update a quote area (draft only)
      parameters:
        - $ref: '#/components/parameters/CustomerId'
        - $ref: '#/components/parameters/QuoteId'
        - $ref: '#/components/parameters/AreaId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateQuoteAreaRequest'
      responses:
        '200':
          description: Area updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QuoteArea'
        '400':
          description: Invalid request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Area, quote, or customer not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Quote is not in draft status.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    delete:
      operationId: removeQuoteArea
      summary: Remove a quote area (draft only)
      parameters:
        - $ref: '#/components/parameters/CustomerId'
        - $ref: '#/components/parameters/QuoteId'
        - $ref: '#/components/parameters/AreaId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ActorRequest'
      responses:
        '200':
          description: Area removed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QuoteArea'
        '400':
          description: Invalid request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Area, quote, or customer not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Area has line items or quote is not in draft status.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /customers/{customerId}/quotes/{quoteId}/areas/{areaId}/pieces:
    get:
      operationId: listQuoteAreaPieces
      summary: List counter pieces for a quote area
      parameters:
        - $ref: '#/components/parameters/CustomerId'
        - $ref: '#/components/parameters/QuoteId'
        - $ref: '#/components/parameters/AreaId'
      responses:
        '200':
          description: Counter pieces returned.
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/CounterPiece'
        '404':
          description: Area, quote, or customer not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      operationId: createQuoteAreaPiece
      summary: Add a counter piece to a quote area (draft only)
      parameters:
        - $ref: '#/components/parameters/CustomerId'
        - $ref: '#/components/parameters/QuoteId'
        - $ref: '#/components/parameters/AreaId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCounterPieceRequest'
      responses:
        '201':
          description: Counter piece added.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CounterPiece'
        '400':
          description: Invalid request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Area, quote, or customer not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Quote is not in draft status.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /customers/{customerId}/quotes/{quoteId}/areas/{areaId}/pieces/{id}:
    patch:
      operationId: updateQuoteAreaPiece
      summary: Update a counter piece (draft only)
      parameters:
        - $ref: '#/components/parameters/CustomerId'
        - $ref: '#/components/parameters/QuoteId'
        - $ref: '#/components/parameters/AreaId'
        - $ref: '#/components/parameters/PieceId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateCounterPieceRequest'
      responses:
        '200':
          description: Counter piece updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CounterPiece'
        '400':
          description: Invalid request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Counter piece, area, quote, or customer not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Quote is not in draft status.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      operationId: removeQuoteAreaPiece
      summary: Remove a counter piece (draft only)
      parameters:
        - $ref: '#/components/parameters/CustomerId'
        - $ref: '#/components/parameters/QuoteId'
        - $ref: '#/components/parameters/AreaId'
        - $ref: '#/components/parameters/PieceId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ActorRequest'
      responses:
        '200':
          description: Counter piece removed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CounterPiece'
        '400':
          description: Invalid request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Counter piece, area, quote, or customer not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Quote is not in draft status.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /customers/{customerId}/quotes/{quoteId}/areas/{areaId}/edges:
    get:
      operationId: listQuoteAreaEdges
      summary: List edge segments for a quote area
      parameters:
        - $ref: '#/components/parameters/CustomerId'
        - $ref: '#/components/parameters/QuoteId'
        - $ref: '#/components/parameters/AreaId'
      responses:
        '200':
          description: Edge segments returned.
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/EdgeSegment'
        '404':
          description: Area, quote, or customer not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      operationId: createQuoteAreaEdge
      summary: Add an edge segment to a quote area (draft only)
      parameters:
        - $ref: '#/components/parameters/CustomerId'
        - $ref: '#/components/parameters/QuoteId'
        - $ref: '#/components/parameters/AreaId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateEdgeSegmentRequest'
      responses:
        '201':
          description: Edge segment added.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EdgeSegment'
        '400':
          description: Invalid request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Area, quote, or customer not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Quote is not in draft status.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /customers/{customerId}/quotes/{quoteId}/areas/{areaId}/edges/{id}:
    patch:
      operationId: updateQuoteAreaEdge
      summary: Update an edge segment (draft only)
      parameters:
        - $ref: '#/components/parameters/CustomerId'
        - $ref: '#/components/parameters/QuoteId'
        - $ref: '#/components/parameters/AreaId'
        - $ref: '#/components/parameters/EdgeId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateEdgeSegmentRequest'
      responses:
        '200':
          description: Edge segment updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EdgeSegment'
        '400':
          description: Invalid request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Edge segment, area, quote, or customer not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Quote is not in draft status.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      operationId: removeQuoteAreaEdge
      summary: Remove an edge segment (draft only)
      parameters:
        - $ref: '#/components/parameters/CustomerId'
        - $ref: '#/components/parameters/QuoteId'
        - $ref: '#/components/parameters/AreaId'
        - $ref: '#/components/parameters/EdgeId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ActorRequest'
      responses:
        '200':
          description: Edge segment removed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EdgeSegment'
        '400':
          description: Invalid request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Edge segment, area, quote, or customer not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Quote is not in draft status.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /customers/{customerId}/quotes/{quoteId}/areas/{areaId}/sinks:
    get:
      operationId: listQuoteAreaSinks
      summary: List sink cutouts for a quote area
      parameters:
        - $ref: '#/components/parameters/CustomerId'
        - $ref: '#/components/parameters/QuoteId'
        - $ref: '#/components/parameters/AreaId'
      responses:
        '200':
          description: Sink cutouts returned.
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/SinkCutout'
        '404':
          description: Area, quote, or customer not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      operationId: createQuoteAreaSink
      summary: Add a sink cutout to a quote area (draft only)
      parameters:
        - $ref: '#/components/parameters/CustomerId'
        - $ref: '#/components/parameters/QuoteId'
        - $ref: '#/components/parameters/AreaId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSinkCutoutRequest'
      responses:
        '201':
          description: Sink cutout added.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SinkCutout'
        '400':
          description: Invalid request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Area, quote, or customer not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Quote is not in draft status.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /customers/{customerId}/quotes/{quoteId}/areas/{areaId}/sinks/{id}:
    patch:
      operationId: updateQuoteAreaSink
      summary: Update a sink cutout (draft only)
      parameters:
        - $ref: '#/components/parameters/CustomerId'
        - $ref: '#/components/parameters/QuoteId'
        - $ref: '#/components/parameters/AreaId'
        - $ref: '#/components/parameters/SinkId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateSinkCutoutRequest'
      responses:
        '200':
          description: Sink cutout updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SinkCutout'
        '400':
          description: Invalid request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Sink cutout, area, quote, or customer not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Quote is not in draft status.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      operationId: removeQuoteAreaSink
      summary: Remove a sink cutout (draft only)
      parameters:
        - $ref: '#/components/parameters/CustomerId'
        - $ref: '#/components/parameters/QuoteId'
        - $ref: '#/components/parameters/AreaId'
        - $ref: '#/components/parameters/SinkId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ActorRequest'
      responses:
        '200':
          description: Sink cutout removed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SinkCutout'
        '400':
          description: Invalid request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Sink cutout, area, quote, or customer not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Quote is not in draft status.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...